+++ CSC8503/GameTech/YiEventSystem.h
@@ -23,7 +23,7 @@ public:
 	virtual void	PushEvent(GAME_EVENT_ID id, INT iArg0, INT iArg1);
 	virtual void	PushEvent(GAME_EVENT_ID id, LPCTSTR szArg0, LPCTSTR szArg1, INT iArg2, INT iArg3);
 
-	virtual void	RegisterEventHandle(const std::string& nameEvent, FUNC_EVENT_HANDLE funHandle, UINT uOwnerData = NULL);
+	virtual void	RegisterEventHandle(const std::string& nameEvent, FUNC_EVENT_HANDLE funHandle, DWORD64 uOwnerData = NULL);
 
 	virtual void	ProcessAllEvent();
